STM32 Interfacing with ILI9341 LVGL Integration (Work in progress)
* data length is changed from 32-bit to 16-bit because SPI library also send only 16-bit data so is of no use, unless logic is changed
* major bug fixed in function ILI9341_Send_16BitData
  - logical AND is used instead of AND

* tft_flush_slow function added which used DrawPixel function and hence very slow
* disp_init function initialized the display
* bug fixed in tft_flush function
  - length was in-correct it was twice for 16-bit mode transfer
  - SPI_DATA_TRANSFER_BITS with this macro we can control the transfer mode type, if 8 bit is selected then length must be twice
  - anyhow 8 bit mode is not working, need to investigate but 16-bit mode is working

CubeMx updated to latest version and hence ioc project file also
Some problem which I observed, UART is not working in proteus when PLL is enabled

+++ 03-ILI9341/BSP/ili9341.h
@@ -135,7 +135,7 @@ uint16_t ILI9341_GetHeight( void );
 
 void ILI9341_SendCommand( uint8_t command, uint8_t *data, uint32_t length );
 void ILI9341_SendData( uint8_t *data, uint32_t length );
-void ILI9341_Send_16BitData( uint16_t *data, uint32_t length );
+void ILI9341_Send_16BitData( uint16_t *data, uint16_t length );
 
 void ILI9341_SetWindow( uint16_t x_start, uint16_t y_start, uint16_t x_end, uint16_t y_end );
 void ILI9341_DrawPixel( uint16_t x, uint16_t y, uint16_t color );
